Add mapnik submodule (prebuilt binaries), update vcpkg
One gotcha here, that I haven't sorted out, is the automated update of
vcpkg dependencies. Specifically, libcurl has updated since we last
updated vcpkg, and machines with existing copies of this repo on, are
going to be stuck with their old versions. We might need to build that
into fetch-deps.bat